feat: split task.rs into task/ module (SRP/DDD, closes #150)
Split the monolithic sipag-core/src/task.rs (620 lines) into focused
submodules under sipag-core/src/task/:

- parser.rs   — pure parse_task_content(content, name, status) with zero
                std::fs imports; testable without disk I/O
- storage.rs  — all file operations: read_task_file(), write_task_file(),
                write_tracking_file(), append_ended(), next_filename(),
                list_tasks(), default_sipag_dir()
- naming.rs   — pure slugify() with zero I/O
- mod.rs      — TaskStatus, TaskFile structs + re-exports from submodules

Key refactors:
- parse_task_file(path) split into parse_task_content (pure) +
  read_task_file (I/O wrapper); callers updated
- Timestamps made injectable: write_task_file(), write_tracking_file(),
  and append_ended() now accept &str instead of calling Utc::now()
  internally; callers supply timestamps
- Duplicate init_dirs() removed from task.rs; sipag-core::init::init_dirs
  exported via lib.rs and used throughout
- init_dirs tests relocated to init.rs
- Checklist functions: none present (already removed in #147)
