+++ sipag-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod executor;
+pub mod init;
 pub mod repo;
 pub mod task;
